extract project card logic into hooks
